move load_test_dataset from combined_datasets to test_helpers
