+++ tests/libs/datasets/combined_dataset_utils_test.py
@@ -2,6 +2,7 @@ import pathlib
 
 import pytest
 from covidactnow.datapublic.common_fields import CommonFields
+
 from libs.datasets import AggregationLevel
 from libs.datasets import combined_dataset_utils
 from libs.datasets import combined_datasets
@@ -16,7 +17,7 @@ from tests import test_helpers
 
 def test_persist_and_load_dataset(tmp_path, nyc_fips):
     region = Region.from_fips(nyc_fips)
-    dataset = combined_datasets.load_test_dataset()
+    dataset = test_helpers.load_test_dataset()
     timeseries_nyc = dataset.get_regions_subset([region])
 
     pointer = combined_dataset_utils.persist_dataset(timeseries_nyc, tmp_path)
@@ -31,9 +32,7 @@ def test_persist_and_load_dataset(tmp_path, nyc_fips):
 
 def test_update_and_load(tmp_path: pathlib.Path, nyc_fips, nyc_region):
     # restricting the datasets being persisted to one county to speed up tests a bit.
-    multiregion_timeseries_nyc = combined_datasets.load_test_dataset().get_regions_subset(
-        [nyc_region]
-    )
+    multiregion_timeseries_nyc = test_helpers.load_test_dataset().get_regions_subset([nyc_region])
     one_region_nyc = multiregion_timeseries_nyc.get_one_region(nyc_region)
     assert one_region_nyc.latest[CommonFields.POPULATION] > 1_000_000
     assert one_region_nyc.region.location_id
